+++ dom/bindings/TypedArray.h
@@ -157,7 +157,7 @@ struct TypedArray_base : public SpiderMonkeyInterfaceObjectStorage,
 template <typename T, JSObject* UnwrapArray(JSObject*),
           T* GetData(JSObject*, bool* isShared, const JS::AutoRequireNoGC&),
           void GetLengthAndDataAndSharedness(JSObject*, uint32_t*, bool*, T**),
-          JSObject* CreateNew(JSContext*, uint32_t)>
+          JSObject* CreateNew(JSContext*, size_t)>
 struct TypedArray
     : public TypedArray_base<T, UnwrapArray, GetLengthAndDataAndSharedness> {
  private:
